+++ src/Ext/WarheadType/Detonate.cpp
@@ -21,7 +21,7 @@
 void WarheadTypeExt::ExtData::Detonate(TechnoClass* pOwner, HouseClass* pHouse, BulletExt::ExtData* pBulletExt, CoordStruct coords)
 {
 	auto const pBullet = pBulletExt ? pBulletExt->OwnerObject() : nullptr;
-
+	
 	if (pOwner && pBulletExt)
 	{
 		auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pOwner->GetTechnoType());
@@ -202,6 +202,10 @@ void WarheadTypeExt::ExtData::DetonateOnOneUnit(HouseClass* pHouse, TechnoClass*
 		if (pTarget->Passengers.NumPassengers > 0 && !pTargetTypeExt->NoManualUnload.Get(false))
 			TechnoExt::PassengersTransfer(pTarget, nullptr, true, true);
 	}
+
+	if (this->GarrisonPenetration)
+		this->ApplyGarrisonPenetration(pHouse, pTarget, pOwner, pBulletExt);
+	
 
 #ifdef LOCO_TEST_WARHEADS
 	if (this->InflictLocomotor)
@@ -568,4 +572,77 @@ void WarheadTypeExt::ExtData::ApplyAmmoModifier(TechnoClass* pTarget, HouseClass
 
 	newCurrentAmmo = newCurrentAmmo < 0 ? 0 : newCurrentAmmo;
 	pTarget->Ammo = newCurrentAmmo > maxAmmo ? maxAmmo : newCurrentAmmo;
+}
+
+void WarheadTypeExt::ExtData::ApplyGarrisonPenetration(HouseClass* pInvokerHouse, TechnoClass* pTarget, TechnoClass* pInvoker, BulletExt::ExtData* pBulletExt)
+{
+	auto pBuilding = abstract_cast<BuildingClass*>(pTarget);
+	if (!pBuilding || !pBuilding->Occupants.Count || !pBulletExt)
+		return;
+
+	auto const pTargetTypeExt = TechnoTypeExt::ExtMap.Find(pTarget->GetTechnoType());
+	if (pTargetTypeExt->ImmuneToGarrisonPenetration)
+		return;
+
+	auto const pBullet = pBulletExt->OwnerObject();
+	double multiplierPercentage = GeneralUtils::GetRangedRandomOrSingleValue(this->GarrisonPenetration_DamageMultiplier);
+	int damage = static_cast<int>(std::round(pBullet->WeaponType->Damage * multiplierPercentage));
+	WarheadTypeClass* pWH = this->OwnerObject();
+
+	if (multiplierPercentage <= 0.0)
+		return;
+
+	int occupantIndex = this->GarrisonPenetration_RandomTarget ? ScenarioClass::Instance->Random.RandomRanged(0, pBuilding->Occupants.Count - 1) : -1;
+
+	auto DoDamage = [=](InfantryClass* pPassenger, int damage)
+	{
+		auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pPassenger->GetTechnoType());
+		if (pTypeExt->ImmuneToGarrisonPenetration)
+			return;
+
+		damage = MapClass::GetTotalDamage(damage, pWH, pPassenger->Type->Armor, 0);
+
+		pPassenger->Health = pPassenger->Health - damage;
+		pPassenger->Health = pPassenger->Health < 0 ? 0 : pPassenger->Health;
+		pPassenger->IsAlive = !pPassenger->Health ? false : pPassenger->IsAlive;
+
+		if (!pPassenger->Health)
+		{
+			if (pPassenger->Type->VoiceDie.Count)
+			{
+				// Play a new death sound.
+				int soundIndex = pPassenger->Type->VoiceDie[Randomizer::Global->Random() % pPassenger->Type->VoiceDie.Count];
+				VocClass::PlayAt(soundIndex, pInvoker->Location);
+			}
+			else if (pPassenger->Type->DieSound.Count)
+			{
+				// Play a new fallback death sound.
+				int soundIndex = pPassenger->Type->DieSound[Randomizer::Global->Random() % pPassenger->Type->DieSound.Count];
+				VocClass::PlayAt(soundIndex, pInvoker->Location);
+			}
+
+			pInvoker->KillPassengers(pPassenger);
+			pInvoker->RegisterDestruction(pPassenger);
+			pPassenger->UnInit();
+		}
+	};
+
+	// Victim's death sound
+	if (occupantIndex < 0)
+	{
+		for (int i = 0; i < pBuilding->Occupants.Count; i++)
+		{
+			InfantryClass* pPassenger = pBuilding->Occupants.GetItem(i);
+			DoDamage(pPassenger, damage);
+		}
+	}
+	else
+	{
+		InfantryClass* pPassenger = pBuilding->Occupants.GetItem(occupantIndex);
+		DoDamage(pPassenger, damage);
+	}
+
+	// Building fully cleaned!
+	if (!pBuilding->Occupants.Count && GarrisonPenetration_CleanSound.isset())
+		VocClass::PlayAt(GarrisonPenetration_CleanSound.Get(), pInvoker->Location);
 }